Simplified protocol configuration files and allowed easy use of custom protocols.
The procedures in gtx_gen.tcl are now more generic and take a protocol config file
as an argument in addition to settings such as 8b10b and GTREFCLK1. This allows projects
using Bedrock as a submodule to specify their own GTX protocols and still leverage the
auto-generation of the qgtx_wrap.v wrapper module.