Revised way of maintaining two types of interfaces for sf_user.
Instead of using `defines, the parameter DPRAM has been moved out of sf_user.
When newad.py is available, it can be used to auto-generate the DPRAM outside
of sf_user. Otherwise, one of the two wrappers in sf_user_wrap.v can be used,
which provide two mechanisms of storing Digaree's parameters: a DPRAM and a
parallel register bank.