Fix user reported issue where an unavailable gtrefclk0 would stop GTP PLLs from locking
The correct gtrefclk will now be assigned to each PLL reset rail. Note that both PLLs
must always be provided with a running gtrefclk to achieve system lock, even if only one
PLL is being used. It is valid to connect the same refclk to both PLLs in these cases.