Enhancing ff_driver.v so it can support FF tables of varying sizes.
FF table size is retrieved from the first memory entry. Play-back
terminates when FF table is read-out but cavity modeling remains
enabled so that it can follow natural cavity response beyond the
FF table.

Being able to specify the size of the FF table allows for higher
pulse repetition rates, as the entire memory need not be fully consumed.