fpga_family: iverilog gets confused when 2+ modules are in the same file
When using the -M option to get the file dependencies,
iverilog will try to satisfy the dependency by looking
at a file with the same name as the missing module and
the module must be the first one in the file. If these
two conditions are not true, iverlog will give you an
error.