.gitlab: fix missing module litex.gen.genlib when using litescope
This was due to the fact that litex_setup.py is using
litescope master branch. Because commit litescope@4d2df690af
stared to use litex.gen.genlib.misc, instead of migen.genlib.misc,
litex also needs to be updated. However, we are stuck to
litex@70c5be6fb8 for now.