new file dsp/freq_gcount.v
Generalize freq_count to freq_gcount, adding a gate in the unknown-frequency clock domain. freq_count stays as a wrapper around freq_gcount that sets the new g_in input to 1. It's a known defect in the Verilog language (designed in the 1990s) that inputs can't be given a default value.

Note that this is a breaking change -- not on the Verilog side, but on the build system side, for projects that use the freq_count module and have static dependency lists. They must add $(BEDROCK)/dsp/freq_gcount.v as a dependency. Projects that autodetect dependencies with e.g., iverilog -y foo -M will continue to work without effort.

This change updates four such static dependency lists within bedrock itself.