EVG_EVR: add more accurate time-of-day receiver
This module measures the clock rate constantly to
get a more accurate increment for the 32-bit nanoseconds
part. This means that the nanoseconds register is not
the old "number of ticks since PPS" anymore, but a fraction
of a second.

To convert from fraction to nanoseconds, you have to divide
the fraction register with the factor 2^32/1e9 (=4.294967296).